fix infinite loop occurence on pile has no card status
userActions.js
pick_field_card checks whether the chosen pile is empty or not.
If pile is empty, then callback an error

actions.js
payRent is updated, if callback on pick_valuable_card returns an error, then the user must be trying to avoid paying the rent.
Error will not be passed to callback. User will be triggered to pay the rent.

playActionCard/index.js
The card played will still be counted regardless whether the user is playing the card properly or not.