Update Database Schema
If you have a created a db already, I'd suggest dropping that database because the new schema updates an enum. And you cannot rollback an enum. You need to either drop the enum manually or just drop the whole database, and I think the later would be quicker and easier imo.